+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Handles an emtpy tensor for convolution (issue #384)
 - When `coordinates` is provided in `MinkowskiToSparseTensor`, `remove_zeros` will be ignored (issue #387)
 - Pybind equality error from the package (issue #414)
+- Fix undefined coordinate merge for multiple coordinate unions
 
 ## [0.5.4]